backend/pyproject.toml
[project]
name = "recipe-kg"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-frontmatter",
    "pydantic",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
pythonpath = [".."]
